# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Guardian Server API
  version: 1.0.0
  description: |
    Guardian 后端 REST API 文档。
    SQLite 持久化 admins/teachers，rooms/clients 全量在内存。
    WebSocket 协议见 `docs/asyncapi.yaml`。
...
servers:
  - url: http://{host}:3847
    variables:
      host:
        default: localhost
...
paths:
  /api/admin/login:
    post:
      tags: [Admin]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/teachers:
    get:
      tags: [Admin]
      summary: 获取教师列表
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, teachers]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  teachers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Teacher'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin]
      summary: 创建教师
      security:
        - XTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, teacher]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  teacher:
                    $ref: '#/components/schemas/Teacher'
        '400':
          description: 参数错误或工号重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/teachers/{id}:
    parameters:
      - $ref: '#/components/parameters/TeacherId'
    get:
      tags: [Admin]
      summary: 获取教师详情
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, teacher]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  teacher:
                    $ref: '#/components/schemas/Teacher'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin]
      summary: 更新教师
      security:
        - XTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, teacher]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  teacher:
                    $ref: '#/components/schemas/Teacher'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin]
      summary: 删除教师
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkOnly'
        '400':
          description: 教师名下有房间，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/rooms:
    get:
      tags: [Admin]
      summary: 获取全平台房间列表
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, rooms]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminRoomSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/rooms/{id}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      tags: [Admin]
      summary: 获取任意房间详情（含客户端列表）
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, room]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  room:
                    allOf:
                      - $ref: '#/components/schemas/Room'
                      - type: object
                        required: [clients]
                        properties:
                          clients:
                            type: array
                            items:
                              $ref: '#/components/schemas/ClientInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin]
      summary: 删除任意房间
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkOnly'
        '400':
          description: 房间中仍有在线子机，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/rooms:
    get:
      tags: [Rooms]
      summary: 获取当前教师的房间列表
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, rooms]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeacherRoomSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Rooms]
      summary: 创建房间
      security:
        - XTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, room]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  room:
                    $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/rooms/{id}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      tags: [Rooms]
      summary: 获取房间详情（含客户端列表）
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, room]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  room:
                    allOf:
                      - $ref: '#/components/schemas/Room'
                      - type: object
                        required: [clients]
                        properties:
                          clients:
                            type: array
                            items:
                              $ref: '#/components/schemas/ClientInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Rooms]
      summary: 更新房间（局部）
      security:
        - XTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [ok, room]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  room:
                    $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Rooms]
      summary: 删除房间
      security:
        - XTokenAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkOnly'
        '400':
          description: 房间中仍有在线子机，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    XTokenAuth:
      type: apiKey
      in: header
      name: x-token
  parameters:
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
    TeacherId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未登录或 token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    OkOnly:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [ok, msg]
      properties:
        ok:
          type: boolean
          enum: [false]
        msg:
          type: string
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminLoginResponse:
      type: object
      required: [ok, token, role, username]
      properties:
        ok:
          type: boolean
          enum: [true]
        token:
          type: string
        role:
          type: string
          enum: [admin]
        username:
          type: string
    Teacher:
      type: object
      required: [id, staffId, name, createdAt]
      properties:
        id:
          type: string
        staffId:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    TeacherCreateRequest:
      type: object
      required: [staffId, name, password]
      properties:
        staffId:
          type: string
        name:
          type: string
        password:
          type: string
    TeacherUpdateRequest:
      type: object
      properties:
        staffId:
          type: string
        name:
          type: string
        password:
          type: string
...
    ProcessEntry:
      type: object
      properties:
        name:
          type: string
          description: 进程名，如 "notepad.exe"
        path:
          type: string
          description: 可选完整路径
        description:
          type: string
        enabled:
          type: boolean
...
    UrlRule:
      type: object
      properties:
        pattern:
          type: string
        type:
          type: string
          enum: [allow, deny]
        description:
          type: string
    GuardConfig:
      type: object
      required: [checkInterval, notifyOnly, autoStartGuard]
      properties:
        checkInterval:
          type: integer
        notifyOnly:
          type: boolean
        autoStartGuard:
          type: boolean
    ScheduleConfig:
      type: object
      required: [autoMode, gracePeriod, allowLateJoin]
      properties:
        autoMode:
          type: boolean
        gracePeriod:
          type: integer
        allowLateJoin:
          type: boolean
    ViolationsConfig:
      type: object
      required: [maxAllowed]
      properties:
        maxAllowed:
          type: integer
    WhitelistItem:
      oneOf:
        - type: string
          description: 纯字符串，如 "notepad.exe"
        - $ref: '#/components/schemas/ProcessEntry'
    RoomWhitelist:
      type: object
      required: [processes, browsers, urls]
      properties:
        processes:
          type: array
          description: 允许进程名，每项可为纯字符串或对象
          items:
            $ref: '#/components/schemas/WhitelistItem'
        browsers:
          type: array
          description: 允许浏览器名，每项可为纯字符串或对象
          items:
            $ref: '#/components/schemas/WhitelistItem'
        urls:
          type: array
          description: 允许域名，每项可为纯字符串或 UrlRule 对象
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/UrlRule'
    Student:
      type: object
      required: [studentId, name]
      properties:
        studentId:
          type: string
        name:
          type: string
    ViolationEntry:
      type: object
      properties:
        pid:
          type: integer
        name:
          type: string
        title:
          type: string
        time:
          type: string
        action:
          type: string
    Room:
      type: object
      required:
        - id
        - roomName
        - joinCode
        - teacherId
        - createdAt
        - guard
        - schedule
        - whitelist
        - violations
        - students
      properties:
        id:
          type: string
        roomName:
          type: string
        joinCode:
          type: string
        teacherId:
          type: string
        createdAt:
          type: string
          format: date-time
        guard:
          $ref: '#/components/schemas/GuardConfig'
        schedule:
          $ref: '#/components/schemas/ScheduleConfig'
        whitelist:
          $ref: '#/components/schemas/RoomWhitelist'
        violations:
          $ref: '#/components/schemas/ViolationsConfig'
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
    TeacherRoomSummary:
      allOf:
        - $ref: '#/components/schemas/Room'
        - type: object
          required: [studentCount, onlineCount]
          properties:
            studentCount:
              type: integer
            onlineCount:
              type: integer
    AdminRoomSummary:
      allOf:
        - $ref: '#/components/schemas/TeacherRoomSummary'
        - type: object
          properties:
            teacherName:
              type: string
    ClientInfo:
      type: object
      required:
        - clientId
        - studentId
        - studentName
        - ip
        - hostname
        - online
        - guardActive
        - processCount
        - lastSeen
        - violations
      properties:
        clientId:
          type: string
        studentId:
          type: string
          nullable: true
        studentName:
          type: string
          nullable: true
        ip:
          type: string
        hostname:
          type: string
        online:
          type: boolean
        guardActive:
          type: boolean
        processCount:
          type: integer
        lastSeen:
          type: integer
          format: int64
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ViolationEntry'
    RoomCreateRequest:
      type: object
      required: [roomName]
      properties:
        roomName:
          type: string
        guard:
          $ref: '#/components/schemas/GuardConfig'
        schedule:
          $ref: '#/components/schemas/ScheduleConfig'
        whitelist:
          $ref: '#/components/schemas/RoomWhitelist'
        violations:
          $ref: '#/components/schemas/ViolationsConfig'
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
    RoomUpdateRequest:
      type: object
      properties:
        roomName:
          type: string
        guard:
          type: object
          additionalProperties: true
        schedule:
          type: object
          additionalProperties: true
        whitelist:
          type: object
          additionalProperties: true
        violations:
          type: object
          additionalProperties: true
...